Switched back to Javalin because of Quarkus issues with Swing
