+++ pom.xml
@@ -19,7 +19,7 @@
     <properties>
         <jdk.version>24</jdk.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <langchain4j.version>1.0.0-beta3</langchain4j.version>
+        <langchain4j.version>1.1.0</langchain4j.version>
         <logback.version>1.5.18</logback.version>
         <junit.version>5.12.1</junit.version>
         <license.maven.plugin.version>5.0.0</license.maven.plugin.version>
@@ -35,12 +35,12 @@
         <opencv.version>4.10.0-1.5.11</opencv.version>
         <openblas.version>0.3.28-1.5.11</openblas.version>
         <a2a.sdk.version>0.2.4-SNAPSHOT</a2a.sdk.version>
-        <quarkus.platform.version>3.22.3</quarkus.platform.version>
+        <!--<quarkus.platform.version>3.22.3</quarkus.platform.version>-->
         <main.class/>
     </properties>
 
     <dependencyManagement>
-        <dependencies>
+        <!--<dependencies>
             <dependency>
                 <groupId>io.quarkus</groupId>
                 <artifactId>quarkus-bom</artifactId>
@@ -48,15 +48,15 @@
                 <type>pom</type>
                 <scope>import</scope>
             </dependency>
-        </dependencies>
+        </dependencies>-->
     </dependencyManagement>
 
     <dependencies>
-        <dependency>
+       <!-- <dependency>
             <groupId>io.a2a.sdk</groupId>
             <artifactId>a2a-java-sdk-server-quarkus</artifactId>
             <version>${a2a.sdk.version}</version>
-        </dependency>
+        </dependency>-->
         <dependency>
             <groupId>io.a2a.sdk</groupId>
             <artifactId>a2a-java-sdk-core</artifactId>
@@ -85,33 +85,33 @@
         </dependency>
         <dependency>
             <groupId>dev.langchain4j</groupId>
-            <artifactId>langchain4j</artifactId>
+            <artifactId>langchain4j-core</artifactId>
             <version>${langchain4j.version}</version>
         </dependency>
         <dependency>
             <groupId>dev.langchain4j</groupId>
             <artifactId>langchain4j-azure-open-ai</artifactId>
-            <version>${langchain4j.version}</version>
+            <version>1.1.0-rc1</version>
         </dependency>
         <dependency>
             <groupId>dev.langchain4j</groupId>
             <artifactId>langchain4j-vertex-ai-gemini</artifactId>
-            <version>${langchain4j.version}</version>
+            <version>1.1.0-beta7</version>
         </dependency>
         <dependency>
             <groupId>dev.langchain4j</groupId>
             <artifactId>langchain4j-google-ai-gemini</artifactId>
-            <version>${langchain4j.version}</version>
+            <version>1.1.0-rc1</version>
         </dependency>
         <dependency>
             <groupId>dev.langchain4j</groupId>
             <artifactId>langchain4j-chroma</artifactId>
-            <version>${langchain4j.version}</version>
+            <version>1.1.0-beta7</version>
         </dependency>
         <dependency>
             <groupId>dev.langchain4j</groupId>
             <artifactId>langchain4j-embeddings-all-minilm-l6-v2</artifactId>
-            <version>${langchain4j.version}</version>
+            <version>1.1.0-beta7</version>
         </dependency>
         <dependency>
             <groupId>org.slf4j</groupId>
@@ -151,6 +151,11 @@
             <version>${jackson.version}</version>
         </dependency>
         <dependency>
+            <groupId>com.fasterxml.jackson.datatype</groupId>
+            <artifactId>jackson-datatype-jsr310</artifactId>
+            <version>${jackson.version}</version>
+        </dependency>
+        <dependency>
             <groupId>com.google.guava</groupId>
             <artifactId>guava</artifactId>
             <version>${guava.version}</version>
@@ -161,6 +166,11 @@
                 </exclusion>
             </exclusions>
         </dependency>
+        <!--<dependency>
+            <groupId>io.quarkiverse.logging.logback</groupId>
+            <artifactId>quarkus-logging-logback</artifactId>
+            <version>1.1.2</version>
+        </dependency>-->
     </dependencies>
     <profiles>
         <profile>
@@ -218,11 +228,11 @@
     </profiles>
     <build>
         <plugins>
-            <plugin>
+            <!--<plugin>
                 <groupId>io.quarkus</groupId>
                 <artifactId>quarkus-maven-plugin</artifactId>
                 <version>${quarkus.platform.version}</version>
-                <!--<extensions>true</extensions>-->
+                &lt;!&ndash;<extensions>true</extensions>&ndash;&gt;
                 <executions>
                     <execution>
                         <goals>
@@ -232,20 +242,20 @@
                         </goals>
                     </execution>
                 </executions>
-            </plugin>
+            </plugin>-->
             <plugin>
                 <groupId>org.apache.maven.plugins</groupId>
                 <artifactId>maven-surefire-plugin</artifactId>
                 <version>${maven.surefire.plugin.version}</version>
                 <configuration>
-                    <argLine>
+                   <!-- <argLine>
                         -javaagent:${settings.localRepository}/org/mockito/mockito-core/${mockito.version}/mockito-core-${mockito.version}.jar
                         -Xshare:off
-                    </argLine>
-                    <systemPropertyVariables>
+                    </argLine>-->
+                   <!-- <systemPropertyVariables>
                         <java.util.logging.manager>org.jboss.logmanager.LogManager</java.util.logging.manager>
                         <maven.home>${maven.home}</maven.home>
-                    </systemPropertyVariables>
+                    </systemPropertyVariables>-->
                 </configuration>
             </plugin>
             <plugin>
