+++ src/main/java/io/a2a/server/agentexecution/RequestHandler.java
@@ -1,0 +1,25 @@
+/*
+ * Copyright (c) 2025.
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at http://www.apache.org/licenses/LICENSE-2.0
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ *
+ *  SPDX-License-Identifier: Apache-2.0
+ */
+
+package io.a2a.server.agentexecution;
+
+import io.a2a.spec.AgentCard;
+import io.a2a.spec.JSONRPCResponse;
+import io.a2a.spec.SendMessageRequest;
+
+public interface RequestHandler {
+    JSONRPCResponse onMessageSend(SendMessageRequest request);
+    AgentCard getAgentCard();
+}
