+++ src/main/resources/application.properties
@@ -1,3 +1,4 @@
-quarkus.http.port=9100
+quarkus.http.port=8005
 quarkus.application.name=ui-test-execution-agent
 quarkus.log.level=INFO
+quarkus.awt.headless=false
