+++ src/main/resources/config.properties
@@ -1,13 +1,14 @@
 # Basic agent configuration
 unattended.mode=false
-port=7070
+port=8005
+host=localhost
 test.mode=false
 
 #-----------------------------------------------
 # RAG config
 # Currently only 'chroma' is supported
 vector.db.provider=chroma
-vector.db.url=http://localhost:8000
+vector.db.url=http://localhost:8020
 retriever.top.n=3
 
 #-----------------------------------------------
